backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==21.2.0
pyautogen==0.4.0
openai==1.55.3
httpx==0.27.2
pandas==2.2.2
openpyxl==3.1.5
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.1
